# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 2026 Taiwan Balloon Carnival API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/registrations:
    get:
      operationId: adminListRegistrations
      tags: [admin]
      summary: List all registrations (admin)
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Registration list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/registrations/export:
    get:
      operationId: adminExportRegistrations
      tags: [admin]
      summary: Export registrations as CSV
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/news:
    post:
      operationId: adminCreateNews
      tags: [admin]
      summary: Create a news article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsBody"
      responses:
        "201":
          description: News article created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/news/{id}:
    put:
      operationId: adminUpdateNews
      tags: [admin]
      summary: Update a news article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsBody"
      responses:
        "200":
          description: News article updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: adminDeleteNews
      tags: [admin]
      summary: Delete a news article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/contestants:
    post:
      operationId: adminCreateContestant
      tags: [admin]
      summary: Create a contestant (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContestantBody"
      responses:
        "201":
          description: Contestant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contestant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/contestants/{id}:
    put:
      operationId: adminUpdateContestant
      tags: [admin]
      summary: Update a contestant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContestantBody"
      responses:
        "200":
          description: Contestant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contestant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: adminDeleteContestant
      tags: [admin]
      summary: Delete a contestant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get registration stats per date (admin)
      responses:
        "200":
          description: Stats per date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DateAvailability"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /admin/sales-overview:
    get:
      operationId: adminGetSalesOverview
      tags: [admin]
      summary: Get aggregated ticketing KPIs and trends for the dashboard (admin)
      responses:
        "200":
          description: Sales overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOverview"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Registration:
      type: object
      required: [id, parentName, phone, ticketCount, eventDate, paymentStatus, createdAt]
      properties:
        id:
          type: number
        parentName:
          type: string
        phone:
          type: string
        email:
          type: [string, 'null']
        ticketCount:
          type: number
        eventDate:
          type: string
          format: date
        ticketType:
          type: [string, 'null']
        amount:
          type: [number, 'null']
        paymentMethod:
          type: [string, 'null']
        paymentStatus:
          type: string
        paymentRef:
          type: [string, 'null']
        qrToken:
          type: [string, 'null']
        checkedInAt:
          type: [string, 'null']
          format: date-time
        confirmationEmailSentAt:
          type: [string, 'null']
          format: date-time
        weekReminderSentAt:
          type: [string, 'null']
          format: date-time
        dayReminderSentAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    DateAvailability:
      type: object
      required: [date, totalCapacity, registered, remaining]
      properties:
        date:
          type: string
          format: date
        totalCapacity:
          type: number
        registered:
          type: number
        remaining:
          type: number
    CreateNewsBody:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        summary:
          type: string
    NewsArticle:
      type: object
      required: [id, title, content, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: string
        summary:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateContestantBody:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        imageUrl:
          type: string
        score:
          type: number
    Contestant:
      type: object
      required: [id, name, description, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: [string, 'null']
        score:
          type: [number, 'null']
        createdAt:
          type: string
          format: date-time
...
    SalesTrendPoint:
      type: object
      required: [date, ticketsSold, revenue]
      properties:
        date:
          type: string
          format: date
        ticketsSold:
          type: number
        revenue:
          type: number
    TicketTypeBreakdown:
      type: object
      required: [ticketType, label, ticketsSold, revenue, percentage]
      properties:
        ticketType:
          type: string
          description: Stored ticket_type value (e.g. "single", "combo"). Empty string for legacy rows with no ticket type.
        label:
          type: string
          description: Human-readable label (e.g. "單日票", "兩日套票", "未指定").
        ticketsSold:
          type: number
        revenue:
          type: number
        percentage:
          type: number
    SessionAvailability:
      type: object
      required: [eventDate, label, totalCapacity, registered, remaining, fillPercentage]
      properties:
        eventDate:
          type: string
          format: date
        label:
          type: string
        totalCapacity:
          type: number
        registered:
          type: number
        remaining:
          type: number
        fillPercentage:
          type: number
    SalesOverview:
      type: object
      required:
        - ticketPriceTwd
        - todayTicketsSold
        - todayRevenue
        - totalTicketsSold
        - totalRevenue
        - totalCapacity
        - overallFillPercentage
        - dailySalesTrend
        - ticketTypeBreakdown
        - sessionAvailability
      properties:
        ticketPriceTwd:
          type: number
          description: Reference single-day ticket price in TWD (for display only; revenue is summed from registration.amount).
        todayTicketsSold:
          type: number
          description: Paid tickets created today (by createdAt date in Asia/Taipei).
        todayRevenue:
          type: number
          description: Revenue from paid registrations created today (sum of registration.amount).
        totalTicketsSold:
          type: number
          description: Cumulative paid ticket count.
        totalRevenue:
          type: number
          description: Cumulative revenue from paid registrations (sum of registration.amount).
        totalCapacity:
          type: number
        overallFillPercentage:
          type: number
        dailySalesTrend:
          type: array
          description: Daily ticket sales (grouped by createdAt) for the last 14 days.
          items:
            $ref: '#/components/schemas/SalesTrendPoint'
        ticketTypeBreakdown:
          type: array
          description: Ticket distribution across event sessions (each event date is a ticket type).
          items:
            $ref: '#/components/schemas/TicketTypeBreakdown'
        sessionAvailability:
          type: array
          description: Per-session capacity / remaining seats.
          items:
            $ref: '#/components/schemas/SessionAvailability'